refactor: rename viz.py to visualization.py for better module naming clarity
