+++ src/ravioli/ai/tools/__init__.py
@@ -1,2 +1,2 @@
 from .sql import get_schema, generate_sql
-from .viz import create_viz_payload, VizStrategy
+from .visualization import create_viz_payload, VizStrategy
