+++ src/ravioli/ai/tools/visualization.py
